web/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-storage==2.10.0
psycopg2-binary>=2.9.0
python-dotenv==1.0.0
google-generativeai==0.8.3
Pillow>=10.0.0
requests>=2.31.0